refactor: standardize error handling and unify duplicate code
Error Handling Improvements:
- Fix fatal errors to use SystemExit(1) instead of silent return
- app.py:558: Conflicting flags now exit with error code
- app.py:148: Unknown help command exits with error code
- status.py:255: Missing directory exits with actionable hint
- imports.py:239,314: Failed export scans exit with provider-specific hints
- Improved error messages with actionable guidance

Code Unification:
- Moved resolve_inputs() to cli_common.py (canonical version)
- Removed duplicate definitions from render.py and app.py
- Canonical version uses strict path checking and sorted candidates
- Updated test to expect SystemExit(1) for unknown help commands

All 120 tests passing
